+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Bug fixes and enhancements
 - Fixed ``solve_normal_cg`` when the linear operator is "nonsquare" (does not map to a space of same dimension),
   by Mathieu Blondel.
 - Added :func:`projection_hypercube <jaxopt.projection.projection_hypercube>`, by Mathieu Blondel.
+- Fixed edge case in :class:`jaxopt.Bisection`, by Mathieu Blondel.
 
 Version 0.3.1.
 --------------